requirements.txt
# ============================================================================
# AI Fundamental Lab 1 - Optimization Algorithms
# ============================================================================
# Python version: >= 3.8
# Installation: pip install -r requirements.txt
# ============================================================================

# -------------------------
# Core Scientific Computing
# -------------------------
numpy>=1.21.0,<2.0.0         # Array operations, mathematical functions

# -------------------------
# Data Manipulation
# -------------------------
pandas>=1.3.0,<3.0.0         # Data analysis, CSV/DataFrame operations

# -------------------------
# Visualization
# -------------------------
matplotlib>=3.4.0,<4.0.0     # Plotting library
seaborn>=0.11.0,<1.0.0       # Statistical data visualization

# -------------------------
# Mathematical Functions
# -------------------------
scipy>=1.7.0,<2.0.0          # Scientific computing (used in Cuckoo Search)

# -------------------------
# Configuration Management
# -------------------------
PyYAML>=6.0,<7.0.0           # YAML configuration file parsing

# -------------------------
# Optional Dependencies
# -------------------------
# tqdm>=4.62.0               # Progress bars (not currently used)
# pytest>=7.0.0              # Unit testing (for development)
# black>=22.0.0              # Code formatting (for development)

# ============================================================================
# Tested Versions (Known Working Configuration)
# ============================================================================
# numpy==1.24.3
# pandas==2.0.3
# matplotlib==3.7.2
# seaborn==0.12.2
# scipy==1.11.1
# PyYAML==6.0.1
# ============================================================================
